Fixed compiler error when PS_ON_PIN is disabled
When the the PS_ON_PIN is disabled (i.e. defined as a negative value in
pins.h e.g. there is no AUX power supply; to remove the entry from menu)
the compiler throws 'error: pasting "DIO" and "-" does not give a valid
preprocessing token' because it still checks the SET_OUTPUT(PS_ON_PIN);
function in ui.cpp and throws the error as the PS_ON_PIN is defined as
unsigned char.
Adding the PS_ON_PIN evaluation in the case statement helps to avoid
this problem.