Allow homing of dual Y steppers individually
When using FEATURE_TWO_YSTEPPER, a second endstop can be set up
by setting MIN_HARDWARE_ENDSTOP_Y2 and Y2_MIN_PIN. When these are
set, hitting the first Y endstop causes only the first stepper to
stop. The second endstop must be hit in order for the second
stepper to stop as well. This allows both sides of a Y gantry to
be synced up automatically.