Z-Probe check if probe never triggered and fix secureDistance
I think this was the original intention of secureDistance (?) and I believe this still gives us correct distances for printers that home at z max, since we move to the optimum probing height in LevelingMethod before a probe run.

I added a threshold check if we have remaining steps, which assumes a bed crash since in theory we should never reach that close. 
(Could just check for zero steps, though we might have a odd trigger in the middle of a crash etc.)