TMC2209 SW UART & Short detection
Added the ability to use software serial in order to communicate with TMC2209's. 
Only supported on "SW_CAPABLE_PLATFORM" therefore: AVR, LPC1768, and STM32's. 
Due is excluded here, but shouldn't trigger a compiler error. 
Added a warning for driver shorts. Different message depending on if it was a short to ground or short to another phase. Which phase is also given at the end of the message.