ZProbe - Control menu
This adds the missing control menu for Z Probes. At the moment only the BLTouch has it's own control options. 
Deploy Pin
Stow Pin
Self Test
Alarm Release

Besides that, there's only two control entries so far and that's: run autoleveling, and: probe once. 

When the probe is active another entry will appear giving you the ability to cancel the autolevel/probe. 

For autoleveling, there's a portion of this that's currently missing for Z max homing printers that occurs at the end of a G32 command.  I'm still debating on a proper way of handling this rather than repeating the code again here.