Fix emergency parser potentially wrong buffer size
SERIAL_BUFFER_SIZE for the due is defined as 256 in the arduino platform now it seems. 
Emergency parser didn't match this size and sometimes emergency commands failed to be read by the fw.

SERIAL_IN_BUFFER now checks if it's been manually set, if not it'll try using SERIAL_BUFFER_SIZE if it's defined. Otherwise 128.

I also did this for SERIAL_RX_BUFFER_SIZE defined in the due's HAL in case it's used somewhere in the future.