executeFString gcode serial bug fix (multiple extruders)
If the tool had a start/end script -> executeFString creates a fake gcode object (with a nullptr source) ->
executeGCode then sets the global activesource to this nullptr as a side-effect.  
We tried accessing that nullptr here in prefetch -> crash.
I think we weren't meant to check the activesource here, but rather the source this is called in?