SKR Mini E3's - Move TFT's HWSERIAL2 and VECT_TAB defines to variant.h
Wanted to make it a little easier to get the TFT going. 
They now only need to define BLUETOOTH_SERIAL to 2 in config.h in order to enable it.
Tidying up the .ini file a little. 
Reverted PeripheralPins.cpp to C in order to fix a linker issue with variant directories.